+++ tests/testthat/test-odds_ratio.R
@@ -22,7 +22,24 @@ test_that("or_glm estimates right OR and CI", {
   expect_equal(result, expected, tolerance = 1e-4, check.attributes = FALSE)
 })
 
-test_that("s_odds_ratio estimates right OR and CI", {
+test_that("or_clogit estimates right OR and CI", {
+  data <- data.frame(
+    rsp = as.logical(c(1, 1, 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1, 1, 0, 0)),
+    grp =    letters[c(1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 1, 1, 1, 2, 2, 2, 3, 3, 2, 2)],
+    strata = LETTERS[c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2)]
+  )
+
+  result <- or_clogit(data, conf_level = 0.95)
+  expected <- list(
+    or_ci = list(# from SAS
+      b = c(est = 0.288, lcl = 0.036, ucl = 2.272),
+      c = c(est = 0.780, lcl = 0.075, ucl = 8.146)
+    )
+  )
+  expect_equal(result, expected, tolerance = 1e-3)
+})
+
+test_that("s_odds_ratio estimates right OR and CI (unstratified analysis)", {
 
   data <- data.frame(
     rsp = as.logical(c(1, 1, 0, 1, 0, 0, 1, 1, 0, 0)),
@@ -45,13 +62,41 @@ test_that("s_odds_ratio estimates right OR and CI", {
   expect_equal(result, expected, tolerance = 1e-4)
 })
 
-test_that("estimate_odds_ratio estimates right OR and CI", {
+test_that("s_odds_ratio estimates right OR and CI (stratified analysis)", {
+
+  set.seed(12)
+  dta <- data.frame(
+    rsp = sample(c(TRUE, FALSE), 100, TRUE),
+    grp = factor(rep(c("A", "B"), each = 50), levels = c("B", "A")),
+    strata = factor(sample(c("C", "D"), 100, TRUE))
+  )
+
+  result <- s_odds_ratio(
+    df = subset(dta, grp == "A"),
+    .var = "rsp",
+    .ref_group = subset(dta, grp == "B"),
+    .in_ref_col = FALSE,
+    .df_row = dta,
+    variables = list(arm = "grp", strata = "strata")
+  )
+
+  expected <- list(
+    or_ci = with_label(
+      c(est = 0.76898, lcl = 0.34242, ucl = 1.72692),
+      "Odds Ratio (95% CI)"
+    )
+  )
+  expect_equal(result, expected, tolerance = 1e-4)
+})
+
+test_that("estimate_odds_ratio estimates right OR and CI (unstratified analysis)", {
   data <- data.frame(
     rsp = as.logical(c(1, 1, 0, 1, 0, 0, 1, 1, 0, 0)),
     grp = letters[c(1, 1, 1, 2, 2, 2, 3, 3, 3, 3)]
   )
 
-  result <- split_cols_by(lyt = NULL, var = "grp", ref_group = "a") %>%
+  result <- basic_table() %>%
+    split_cols_by(var = "grp", ref_group = "a") %>%
     estimate_odds_ratio(vars = "rsp") %>%
     build_table(df = data)
 
@@ -62,6 +107,27 @@ test_that("estimate_odds_ratio estimates right OR and CI", {
       "c", "0.5 (0.02 - 11.09)"
     ),
     .Dim = c(2L, 4L)
+  )
+  expect_identical(result_matrix, expected_matrix)
+})
+
+test_that("estimate_odds_ratio estimates right OR and CI (stratified analysis)", {
+  set.seed(12)
+  data <- data.frame(
+    rsp = sample(c(TRUE, FALSE), 100, TRUE),
+    grp = factor(rep(c("A", "B"), each = 50), levels = c("B", "A")),
+    strata = factor(sample(c("C", "D"), 100, TRUE))
+  )
+
+  result <- basic_table() %>%
+    split_cols_by(var = "grp", ref_group = "A") %>%
+    estimate_odds_ratio(vars = "rsp", variables = list(arm = "grp", strata = "strata")) %>%
+    build_table(df = data)
+
+  result_matrix <- to_string_matrix(result)
+  expected_matrix <- structure(
+    c("", "Odds Ratio (95% CI)", "A", "", "B", "1.3 (0.58 - 2.92)"),
+    .Dim = 2:3
   )
   expect_identical(result_matrix, expected_matrix)
 })